Only attempt to kill Chromium process if it hasn't already exited
Previously, we had to manually kill the Chromium task after it had been used. The newer version of PuppeteerSharp seems to handle this better, so we're now finding that the process has already been killed when we get to this point.

+++ PDF2IMG/PageRenderer.cs
@@ -78,7 +78,10 @@ namespace BarnardTech.PDF2IMG
 
             chromeBrowser.Disconnected += (sender, e) =>
             {
-                chromeBrowser.Process.Kill();
+                if (!chromeBrowser.Process.HasExited)
+                {
+                    chromeBrowser.Process.Kill();
+                }
             };
 
             fileResourceHandlerFactory = new FileResourceHandlerFactory("pdfviewer", "host", Directory.GetCurrentDirectory());
